+++ .golangci.yml
@@ -88,7 +88,7 @@ linters:
     - exhaustive
     - exportloopref
     - funlen
-    # - gci
+    - gci
     - gocognit
     - goconst
     - gocritic
