+++ hw12_13_14_15_calendar/internal/server/rpc/server/server.go
@@ -6,13 +6,38 @@ import (
 
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
+	"google.golang.org/protobuf/types/known/timestamppb"
 	interfaces "hw12_13_14_15_calendar/internal/interfaces"
-	catendarrpcapi "hw12_13_14_15_calendar/internal/server/rpc/api"
-	common "hw12_13_14_15_calendar/internal/server/rpc/calendarcommon"
+	models "hw12_13_14_15_calendar/internal/models"
+	calendarrpcapi "hw12_13_14_15_calendar/internal/server/rpc/api"
 )
 
+func Event2PBEvent(event *models.Event) *calendarrpcapi.Event {
+	pbEvent := new(calendarrpcapi.Event)
+	pbEvent.PK = int32(event.PK)
+	pbEvent.Title = event.Title
+	pbEvent.Description = event.Description
+	pbEvent.StartAt = timestamppb.New(event.StartAt)
+	pbEvent.Duration = int32(event.Duration)
+	pbEvent.Owner = event.Owner
+	pbEvent.NotifyEarly = int32(event.NotifyEarly)
+	return pbEvent
+}
+
+func PBEvent2Event(pbEvent *calendarrpcapi.Event) *models.Event {
+	event := new(models.Event)
+	event.PK = int(pbEvent.PK)
+	event.Title = pbEvent.Title
+	event.Description = pbEvent.Description
+	event.StartAt = pbEvent.StartAt.AsTime()
+	event.Duration = int(pbEvent.Duration)
+	event.Owner = pbEvent.Owner
+	event.NotifyEarly = int(pbEvent.NotifyEarly)
+	return event
+}
+
 type RPCServer struct {
-	catendarrpcapi.UnimplementedApplicationServer
+	calendarrpcapi.UnimplementedApplicationServer
 	logger interfaces.Logger
 	app    interfaces.Applicationer
 	server *grpc.Server
@@ -25,53 +50,53 @@ func NewRPCServer(app interfaces.Applicationer, logger interfaces.Logger) *RPCSe
 	return self
 }
 
-func (s *RPCServer) CreateEvent(ctx context.Context, pbEvent *catendarrpcapi.Event) (*catendarrpcapi.Event, error) {
+func (s *RPCServer) CreateEvent(ctx context.Context, pbEvent *calendarrpcapi.Event) (*calendarrpcapi.Event, error) {
 	_ = ctx // TODO: pass to s.app
-	event := common.PBEvent2Event(pbEvent)
+	event := PBEvent2Event(pbEvent)
 	createdEvent, err := s.app.CreateEvent(event)
 	if err != nil {
 		return nil, err
 	}
-	return common.Event2PBEvent(createdEvent), nil
+	return Event2PBEvent(createdEvent), nil
 }
 
-func (s *RPCServer) ReadEvent(ctx context.Context, ident *catendarrpcapi.Id) (*catendarrpcapi.Event, error) {
+func (s *RPCServer) ReadEvent(ctx context.Context, ident *calendarrpcapi.Id) (*calendarrpcapi.Event, error) {
 	_ = ctx // TODO: pass to s.app
 	event, err := s.app.ReadEvent(int(ident.Pk))
 	if err != nil {
 		return nil, err
 	}
-	return common.Event2PBEvent(event), nil
+	return Event2PBEvent(event), nil
 }
 
-func (s *RPCServer) UpdateEvent(ctx context.Context, pbEvent *catendarrpcapi.Event) (*catendarrpcapi.Event, error) {
+func (s *RPCServer) UpdateEvent(ctx context.Context, pbEvent *calendarrpcapi.Event) (*calendarrpcapi.Event, error) {
 	_ = ctx // TODO: pass to s.app
-	event := common.PBEvent2Event(pbEvent)
+	event := PBEvent2Event(pbEvent)
 	updatedEvent, err := s.app.UpdateEvent(event)
 	if err != nil {
 		return nil, err
 	}
-	return common.Event2PBEvent(updatedEvent), nil
+	return Event2PBEvent(updatedEvent), nil
 }
 
-func (s *RPCServer) DeleteEvent(ctx context.Context, pbEvent *catendarrpcapi.Event) (*catendarrpcapi.Event, error) {
+func (s *RPCServer) DeleteEvent(ctx context.Context, pbEvent *calendarrpcapi.Event) (*calendarrpcapi.Event, error) {
 	_ = ctx // TODO: pass to s.app
-	event := common.PBEvent2Event(pbEvent)
+	event := PBEvent2Event(pbEvent)
 	deletedEvent, err := s.app.DeleteEvent(event)
 	if err != nil {
 		return nil, err
 	}
-	return common.Event2PBEvent(deletedEvent), nil
+	return Event2PBEvent(deletedEvent), nil
 }
 
-func (s *RPCServer) ListEvents(_ *emptypb.Empty, stream catendarrpcapi.Application_ListEventsServer) error {
+func (s *RPCServer) ListEvents(_ *emptypb.Empty, stream calendarrpcapi.Application_ListEventsServer) error {
 	events, err := s.app.ListEvents()
 	if err != nil {
 		return err
 	}
 	for _, event := range events {
 		tmp := event
-		pbEvent := common.Event2PBEvent(&tmp)
+		pbEvent := Event2PBEvent(&tmp)
 		if err := stream.Send(pbEvent); err != nil {
 			return err
 		}
@@ -133,7 +158,7 @@ func (s *RPCServer) Start(address string) error {
 		grpc.UnaryInterceptor(LoggedUnaryInterceptor(s.logger)),
 		grpc.StreamInterceptor(LoggedStreamInterceptor(s.logger)),
 	)
-	catendarrpcapi.RegisterApplicationServer(gRPCServer, s)
+	calendarrpcapi.RegisterApplicationServer(gRPCServer, s)
 	s.server = gRPCServer
 	s.logger.Info("GRPCServer.Start()")
 	return s.server.Serve(lis)
